Simplify data splitting
