+++ nevergrad/instrumentation/test_multivariables.py
@@ -3,12 +3,29 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from typing import Any, Tuple, Dict
+from typing import Any, Tuple, Dict, List
 import numpy as np
 from ..common import testing
 from . import variables as var
 from . import multivariables as mvar
 from .core import Variable
+
+
+@testing.parametrized(
+    empty=([], [], [])
+)
+def test_split_data(tokens: List[Variable], data: List[float], expected: List[List[float]]) -> None:
+    instru = mvar.Instrumentation(*tokens)
+    output = instru._split_data(np.array(data))
+    testing.printed_assert_equal(output, expected)
+
+
+def test_instrumentation_data_to_arguments() -> None:
+    tokens = [var.SoftmaxCategorical(list(range(5))), var.Gaussian(3, 4)]
+    instru = mvar.Instrumentation(*tokens)
+    values = instru.data_to_arguments([0, 200, 0, 0, 0, 2])[0]
+    np.testing.assert_equal(values, [1, 11])
+    np.testing.assert_raises(ValueError, instru.data_to_arguments, tokens, [0, 200, 0, 0, 0, 2, 3])
 
 
 def test_instrumentation() -> None:
@@ -54,7 +71,7 @@ def test_instrumentation_copy() -> None:
 
 def test_instrumentation_split() -> None:
     instru = mvar.Instrumentation(var.Gaussian(0, 1), 3, b=var.SoftmaxCategorical([0, 1, 2, 3]), a=var.OrderedDiscrete([0, 1, 2, 3]))
-    splitted = instru.split_data([0, 1, 2, 3, 4, 5])
+    splitted = instru._split_data(np.array([0, 1, 2, 3, 4, 5]))
     np.testing.assert_equal([x.tolist() for x in splitted], [[0], [], [1], [2, 3, 4, 5]])  # order of kwargs is alphabetical
